fix(guard): parse proxy Grok hooks and revert expired Watch
